Move QuickStartTaskCardAdapter to upper level
